kmod: upgrade from 31 to 32
kmod now installs symlinks in 'make install'.
https://git.kernel.org/pub/scm/utils/kernel/kmod/kmod.git/commit/?
id=e98cef6f3f8cd6f8bfb26d147b2c209297453cca

Changes are made in do_install to ensure that we can to continue to
use sbin. This is because our kernel has: CONFIG_MODPROBE_PATH='/sbin/modprobe'.
We'll need to sync our kmod path settings with our kernel config. Otherwise,
things would fail. For example, libnl's some ptest cases fail.

+++ meta/recipes-kernel/kmod/kmod_32.bb
@@ -15,7 +15,7 @@ LIC_FILES_CHKSUM = "file://COPYING;md5=a6f89e2100d9b6cdffcea4f398e37343 \
                    "
 inherit autotools bash-completion gtk-doc pkgconfig manpages update-alternatives
 
-SRCREV = "aff617ea871d0568cc491bd116c0be1e857463bb"
+SRCREV = "41faa59711742c1476d59985011ee0f27ed91d30"
 
 SRC_URI = "git://git.kernel.org/pub/scm/utils/kernel/kmod/kmod.git;branch=master;protocol=https \
            file://depmod-search.conf \
@@ -52,8 +52,9 @@ do_install:append () {
         install -dm755 ${D}${base_bindir}
         install -dm755 ${D}${base_sbindir}
         # add symlinks to kmod
-        ln -rs ${D}${base_bindir}/kmod ${D}${base_bindir}/lsmod
+        [ -e ${D}${base_bindir}/lsmod ] || ln -rs ${D}${base_bindir}/kmod ${D}${base_bindir}/lsmod
         for tool in insmod rmmod depmod modinfo modprobe; do
+                rm -f ${D}${base_bindir}/${tool}
                 ln -rs ${D}${base_bindir}/kmod ${D}${base_sbindir}/${tool}
         done
         # configuration directories
